text_generator/pyproject.toml
[project]
name = "text-generator"
version = "0.0"
description = "Pipeline for generate synthetic texts"
authors = [
    {name = "adrian-gude", email = "adrianlopezgude@gmail.com"},
    {name = "dunque", email = "roi.santos.rios@udc.es"}
]
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "transformers (>=4.56.0,<5.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "accelerate (>=1.10.0,<2.0.0)",
    "bitsandbytes (>=0.47.0,<0.48.0)",
    "huggingface-hub[cli] (>=0.34.4,<0.35.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "safetensors (>=0.6.2,<0.7.0)",
    "kernels (>=0.9.0,<0.10.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "stanza (>=1.10.1,<2.0.0)",
    "deepspeed (>=0.17.5,<0.18.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
    "openai (>=1.108.2,<2.0.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
]
requires-python = ">=3.12, <3.14"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
pytest-cov = "^6.2.1"
